Update and rename final_assignment to final_assignment.py
Added new version of the code. This is skeleton code that lacks any data acquisition but does store the data in a .h5 file. It does so via appending at each step rather than simply creating a new file at the end of the code meaning that the code can crash without the data being compromised.